Modificacoes da classe Partida Service. Devolvendo o Status Code dos metodos deletePartida e cadastrarPartida. Incluindo uma Exception no metodo alteraPartida

+++ partida-futebol-api/src/main/java/br/com/meli/partidafutebolapi/service/PartidaService.java
@@ -2,10 +2,14 @@ package br.com.meli.partidafutebolapi.service;
 
 import br.com.meli.partidafutebolapi.dto.AlteraPartidaDto;
 import br.com.meli.partidafutebolapi.dto.CadastroPartidaDto;
-import br.com.meli.partidafutebolapi.model.PartidaModel;
+import br.com.meli.partidafutebolapi.model.Partida;
 import br.com.meli.partidafutebolapi.repository.PartidaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -15,28 +19,31 @@ import java.util.stream.Collectors;
 public class PartidaService {
     @Autowired
     public PartidaRepository repository;
-    public void cadastrarPartida(CadastroPartidaDto cadastro){
-        PartidaModel partida = new PartidaModel();
-        partida.setId(cadastro.getId());
+    public ResponseEntity cadastrarPartida(CadastroPartidaDto cadastro){
+        Partida partida = new Partida();
+
         partida.setClubeMandante(cadastro.getClubeMandante());
         partida.setClubeVisitante(cadastro.getClubeVisitante());
         partida.setGoalsMandante(cadastro.getGoalsMandante());
         partida.setGoalsVisitante(cadastro.getGoalsVisitante());
         partida.setNomeEstadio(cadastro.getEstadio());
         partida.setDataHoraPartida(cadastro.getDataHoraPartida());
+
         repository.save(partida);
+
+        return ResponseEntity.status(201).build();
     }
 
     public List<CadastroPartidaDto> getAllPartidas(){
-        List<PartidaModel> partidas = repository.findAll();
+        List<Partida> partidas = repository.findAll();
         return convertToListPartidasDto(partidas);
     }
 
     public CadastroPartidaDto alteraPartida(AlteraPartidaDto alteraPartida, Long id) {
 
-        Optional<PartidaModel> partidas = repository.findById(id);
+        Optional<Partida> partidas = repository.findById(id);
         if(partidas.isPresent()){
-            PartidaModel partidaAlterada = partidas.get();
+            Partida partidaAlterada = partidas.get();
             partidaAlterada.setClubeMandante(alteraPartida.getClubeMandante());
             partidaAlterada.setClubeVisitante(alteraPartida.getClubeVisitante());
             partidaAlterada.setGoalsMandante(alteraPartida.getGoalsMandante());
@@ -46,19 +53,31 @@ public class PartidaService {
             repository.save(partidaAlterada);
 
             return convertToPartidaAlteradaDto(partidaAlterada);
+
+        }else{
+
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
-
-        return null;
     }
 
-    public void deletePartida(Long id){
-        repository.deleteById(id);
+    public ResponseEntity deletePartida(Long id){
+            Optional<Partida> partidas = repository.findById(id);
+
+            if(partidas.isEmpty()){
+                return ResponseEntity.notFound().build();
+
+            }else{
+
+                 repository.deleteById(id);
+                 return ResponseEntity.noContent().build();
+            }
+
     }
 
-    private static List<CadastroPartidaDto> convertToListPartidasDto(List<PartidaModel> partidas) {
+    private static List<CadastroPartidaDto> convertToListPartidasDto(List<Partida> partidas) {
         return partidas.stream().map(CadastroPartidaDto::new).collect(Collectors.toList());
     }
-    private CadastroPartidaDto convertToPartidaAlteradaDto(PartidaModel partida) {
+    private CadastroPartidaDto convertToPartidaAlteradaDto(Partida partida) {
         CadastroPartidaDto cadastro = new CadastroPartidaDto();
 
         cadastro.setId(partida.getId());
